backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
aiofiles==23.2.1
pandas==2.1.3
scikit-learn==1.3.2
pillow==10.1.0
email-validator>=2.0.0